refactor: Update readLatestLadduAllotment to readLatestLadduSession
